Migrate Mutation.updateOIDCIntegration
